% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sessionInfo.R
\name{compare_si}
\alias{compare_si}
\title{Compare two \code{sessionInfo} objects}
\usage{
compare_si(si1, si2)
}
\arguments{
\item{si1}{A list, corresponding to the first object, generated by \code{\link[=parse_si]{parse_si()}}}

\item{si2}{A list, corresponding to the second object, generated by \code{\link[=parse_si]{parse_si()}}}
}
\value{
A data frame, with the packages, their versions, and the result of the
comparison of their versions
}
\description{
Compare two \code{sessionInfo} objects, parsed by \code{\link[=parse_si]{parse_si()}}
}
\examples{
si1_output <- system.file("extdata", "si1.txt", package = "sessionDiffo")
si2_output <- system.file("extdata", "si2.txt", package = "sessionDiffo")
si1 <- parse_si(si1_output)
si2 <- parse_si(si2_output)
comparison_df <- compare_si(si1, si2)
}
