% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessionDiffo-pkg.R
\docType{package}
\name{sessionDiffo-pkg}
\alias{sessionDiffo-pkg}
\title{\code{sessionDiffo}}
\description{
A tiny utility to compare (\code{diff}) the output of two different Session Info
(via \code{sessionInfo()}, \code{sessioninfo::session_info()}) calls
}
