% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssd.R
\name{ssd}
\alias{ssd}
\title{Shiny up sessionDiffo}
\usage{
ssd()
}
\value{
A Shiny app
}
\description{
A mini Shiny app for comparing \code{sessionInfo} outputs
}
\examples{
if(interactive())
  ssd()
# and later upload the files/copy-paste the output in the editors directly
}
